Add JS to move footnotes to References section
Introduced assets/js/footnotes.js to automatically move footnotes from the bottom of the page to the References section if present. Updated scripts.html to include the new script and added related CSS in style.scss to control footnote visibility and styling. This improves the organization and presentation of footnotes on the site.